move action from the subfolder to root
