+++ internal/agent/agent.go
@@ -7,10 +7,10 @@ import (
 	"os"
 	"path/filepath"
 
-	e2e "github.com/newrelic/newrelic-integration-e2e-action/newrelic-integration-e2e/internal"
-	"github.com/newrelic/newrelic-integration-e2e-action/newrelic-integration-e2e/internal/spec"
-	"github.com/newrelic/newrelic-integration-e2e-action/newrelic-integration-e2e/pkg/dockercompose"
-	"github.com/newrelic/newrelic-integration-e2e-action/newrelic-integration-e2e/pkg/oshelper"
+	e2e "github.com/newrelic/newrelic-integration-e2e-action/internal"
+	"github.com/newrelic/newrelic-integration-e2e-action/internal/spec"
+	"github.com/newrelic/newrelic-integration-e2e-action/pkg/dockercompose"
+	"github.com/newrelic/newrelic-integration-e2e-action/pkg/oshelper"
 	"github.com/sirupsen/logrus"
 	yaml "gopkg.in/yaml.v3"
 )
