+++ internal/agent/agent_test.go
@@ -8,9 +8,9 @@ import (
 	"strings"
 	"testing"
 
-	e2e "github.com/newrelic/newrelic-integration-e2e-action/newrelic-integration-e2e/internal"
-	"github.com/newrelic/newrelic-integration-e2e-action/newrelic-integration-e2e/internal/agent"
-	"github.com/newrelic/newrelic-integration-e2e-action/newrelic-integration-e2e/pkg/oshelper"
+	e2e "github.com/newrelic/newrelic-integration-e2e-action/internal"
+	"github.com/newrelic/newrelic-integration-e2e-action/internal/agent"
+	"github.com/newrelic/newrelic-integration-e2e-action/pkg/oshelper"
 	"github.com/stretchr/testify/require"
 )
