+++ internal/agent/resources/docker-compose.yml
