+++ internal/newrelic/client_test.go
