+++ internal/runtime/entities_tester.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/sirupsen/logrus"
 
-	"github.com/newrelic/newrelic-integration-e2e-action/newrelic-integration-e2e/internal/newrelic"
-	"github.com/newrelic/newrelic-integration-e2e-action/newrelic-integration-e2e/internal/spec"
+	"github.com/newrelic/newrelic-integration-e2e-action/internal/newrelic"
+	"github.com/newrelic/newrelic-integration-e2e-action/internal/spec"
 )
 
 type EntitiesTester struct {
