+++ internal/spec/definition_test.go
