+++ test/testdata/kafka/agent_dir/docker-compose.yml
