+++ test/testdata/kafka/agent_dir/newrelic-infra-agent/Dockerfile
