+++ test/testdata/powerdns/deps/pdns-authoritative/fixdsrrs.sh
