+++ test/testdata/powerdns/deps/pdns-authoritative/pdns.local.gmysql.conf
