+++ test/testdata/powerdns/deps/pdns-authoritative/pdns.preference
