+++ test/testdata/powerdns/deps/pdns-authoritative/schema.mysql.sql
