+++ test/testdata/powerdns/deps/pdns-authoritative/secallzones.sh
