+++ test/testdata/powerdns/deps/pdns-authoritative/start.sh
